requirements.txt
absl-py==1.4.0
aiohttp==3.8.6
aiosignal==1.2.0
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
apturl==0.5.2
asn1crypto==0.24.0
astor==0.8.1
async-timeout==4.0.2
asynctest==0.13.0
attrdict==2.0.1
attrs==22.2.0
Babel==2.11.0
bce-python-sdk==0.9.7
beautifulsoup4==4.12.3
blinker==1.4
Brlapi==0.6.6
cachetools==4.2.4
cairocffi==1.2.0
CairoSVG==2.7.1
certifi==2024.2.2
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.0.12
click==8.0.4
cloudpickle==2.2.1
cmake==3.28.4
colorama==0.3.7
coloredlogs==15.0.1
commonmark==0.9.1
cryptography==2.1.4
cssselect==1.1.0
cssselect2==0.4.1
cssutils==2.3.1
cupshelpers==1.0
cycler==0.10.0
Cython==0.29.36
dataclasses==0.8
datasets==2.4.0
decorator==4.4.2
defer==1.0.6
defusedxml==0.7.1
detectron2 @ git+https://github.com/facebookresearch/detectron2.git@4aca4bdaa9ad48b8e91d7520e0d0815bb8ca0fb1
dill==0.3.4
distlib==0.3.8
distro==1.9.0
distro-info===0.18ubuntu0.18.04.1
easyocr==1.7.1
einops==0.4.1
et-xmlfile==1.1.0
feedparser==5.2.1
filelock==3.4.1
Flask==2.0.3
Flask-Babel==2.0.0
flatbuffers==24.3.7
frozenlist==1.2.0
fsspec==2022.1.0
future==1.0.0
fvcore==0.1.3.post20210317
gast==0.3.3
gdown==4.7.3
getch==1.0
google-api-core==1.32.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-cloud-vision==1.0.0
googleapis-common-protos==1.56.3
graphsurgeon==0.4.5
grpcio==1.48.2
html5lib==0.999999999
httplib2==0.9.2
huggingface-hub==0.4.0
humanfriendly==10.0
idna==3.6
idna-ssl==1.1.0
imageio==2.15.0
imgaug==0.4.0
importlib-metadata==4.8.3
importlib-resources==5.4.0
iopath==0.1.10
itsdangerous==2.0.1
jetson-stats==4.2.6
Jetson.GPIO==2.0.17
Jinja2==3.0.3
joblib==1.1.1
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.3.1
language-selector==0.1
launchpadlib==1.10.6
layoutparser==0.3.4
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
Levenshtein==0.21.1
linecache2==1.0.0
lmdb==1.4.1
louis==3.5.0
lxml==4.2.1
macaroonbakery==1.1.3
Mako==1.0.7
Markdown==3.3.7
MarkupSafe==2.0.1
matplotlib==3.3.4
mpmath==1.3.0
multidict==5.2.0
multiprocess==0.70.12.2
networkx==2.5.1
ninja==1.11.1.1
numpy==1.19.5
oauth==1.0.1
oauthlib==3.2.2
olefile==0.45.1
omegaconf==2.3.0
onboard==1.4.1
onnx==1.12.0
onnx-simplifier==0.1.9
onnxruntime==1.10.0
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
openpyxl==3.1.2
opt-einsum==3.3.0
optimum==1.1.1
packaging==21.3
paddle-python==0.6.0
paddleocr==2.6.0.1
paddlepaddle-gpu==2.0.0
pandas==1.1.5
pathlib==1.0.1
pbr==3.1.1
pdf2image==1.17.0
pdfminer.six==20211012
pdfplumber==0.6.0
pexpect==4.2.1
Pillow==8.4.0
pkg_resources==0.0.0
platformdirs==2.4.0
portalocker==2.7.0
ppocr-onnx==0.0.3.9
premailer==3.10.0
prettytable==2.5.0
protobuf==3.16.0
pyarrow==6.0.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycairo==1.16.2
pyclipper==1.3.0.post5
pycocotools==2.0.7
pycparser==2.21
pycrypto==2.6.1
pycryptodome==3.20.0
-e git+https://github.com/inducer/pycuda.git@51e956c9c1d3d624055b336d90e8e701cdefad7d#egg=pycuda
pycups==1.9.73
pydot==1.4.2
Pygments==2.14.0
PyGObject==3.26.1
PyJWT==1.5.3
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==3.1.2
pyphen==0.11.0
pyRFC3339==1.0
PySocks==1.7.1
pytesseract==0.3.8
python-apt==1.6.6
python-bidi==0.4.2
python-dateutil==2.9.0.post0
python-debian==0.1.32
python-docx==0.8.11
pytools==2022.1.12
pytz==2024.1
PyWavelets==1.1.1
pyxattr==0.6.0
pyxdg==0.25
PyYAML==6.0.1
rapidfuzz==2.11.1
rarfile==4.1
regex==2023.8.8
reportlab==3.4.0
requests==2.27.1
requests-oauthlib==2.0.0
requests-unixsocket==0.1.5
responses==0.17.0
rich==12.6.0
rsa==4.9
sacremoses==0.0.53
scikit-build==0.16.7
scikit-image==0.17.2
scipy==1.5.4
scour==0.36
SecretStorage==2.3.1
Shapely==1.8.5.post1
simplejson==3.13.2
six==1.16.0
smbus2==0.4.3
soupsieve==2.3.2.post1
ssh-import-id==5.7
subprocess32==3.5.4
sympy==1.9
system-service==0.3
systemd-python==234
table-ocr==0.2.5
tabula==1.0.5
tabula-py==2.3.0
tabulate==0.8.10
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorrt==8.2.1.8
termcolor==1.1.0
tesseract==0.1.3
testresources==2.0.0
tifffile==2020.9.3
tinycss2==1.1.1
tokenizers==0.12.1
torch==1.10.2
torch-vision==0.1.6.dev0
torchvision==0.11.3
tqdm==4.64.1
traceback2==1.4.0
transformers==4.18.0
typing_extensions==4.1.1
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
uff==0.6.9
Unidecode==1.3.8
unittest2==1.1.0
unity-scope-calculator==0.1
unity-scope-chromiumbookmarks==0.1
unity-scope-colourlovers==0.1
unity-scope-devhelp==0.1
unity-scope-firefoxbookmarks==0.1
unity-scope-manpages==0.1
unity-scope-openclipart==0.1
unity-scope-texdoc==0.1
unity-scope-tomboy==0.1
unity-scope-virtualbox==0.1
unity-scope-yelp==0.1
unity-scope-zotero==0.1
urllib3==1.26.18
urwid==2.0.1
virtualenv==20.17.1
visualdl==2.4.2
wadllib==1.3.2
Wand==0.6.13
wcwidth==0.2.13
WeasyPrint==52.5
webencodings==0.5
Werkzeug==2.0.3
xkit==0.0.0
xxhash==3.2.0
yacs==0.1.8
yarl==1.7.2
youtube_dl==2018.3.14
zipp==3.6.0
zope.interface==4.3.2
